Cut the default nodes.json down to theforeman and two agents.
Added Shared Folder attribute for each node.
Fixed the foreman web interface.
The ipaddress in the CentOS VM had a default INPUT REJECT & FORWARD REJECT which blocked as the open INPUT rules that are appended to the table.
The provision script was adding puppet repo references which are no longer needed, CentOS VM has them already.